backend/requirements.txt
alembic==1.16.4
blinker==1.9.0
cachelib==0.13.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cors==1.0.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==25.5.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mercadopago==2.3.0
msgspec==0.19.0
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.4
requests-file==2.1.0
SQLAlchemy==2.0.42
tldextract==5.3.0
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==5.1.1
zope.interface==7.2
